requirements.txt
jupyter
folium
kafka-python
numpy
requests
